ブートローダーにカーネルを起動するプログラムを追加

+++ bootloader/includes/utils.h
@@ -3,5 +3,5 @@
 
 //関数のプロトタイプ宣言
 extern void checkResult(EFI_SYSTEM_TABLE *, EFI_STATUS, CHAR16 *);
-extern void intToChar16(int, CHAR16*, int);
+extern void intToChar16(UINT64, CHAR16*, int);
 extern void halt(void);
